# airflow/api_connexion/openapi/v1.yaml
...
openapi: 3.0.3
info:
  title: "Airflow API (Stable)"
  description: |
    # Overview

    To facilitate management, Apache Airflow supports a range of REST API endpoints across its
    objects.
    This section provides an overview of the API design, methods, and supported use cases.

    Most of the endpoints accept `JSON` as input and return `JSON` responses.
    This means that you must usually add the following headers to your request:
    ```
    Content-type: application/json
    Accept: application/json
    ```

    ## Resources

    The term `resource` refers to a single type of object in the Airflow metadata. An API is broken up by its
    endpoint's corresponding resource.
    The name of a resource is typically plural and expressed in camelCase. Example: `dagRuns`.

    Resource names are used as part of endpoint URLs, as well as in API parameters and responses.

    ## CRUD Operations

    The platform supports **C**reate, **R**ead, **U**pdate, and **D**elete operations on most resources.
    You can review the standards for these operations and their standard parameters below.

    Some endpoints have special behavior as exceptions.

    ### Create

    To create a resource, you typically submit an HTTP `POST` request with the resource's required metadata
    in the request body.
    The response returns a `201 Created` response code upon success with the resource's metadata, including
...

    ### Read

...

...

...

...
    ```
...

...

...

...

...
servers:
...
paths:
...
  /roles:
    get:
      deprecated: true
      summary: List roles
      description: |
        Get a list of roles.

        *This API endpoint is deprecated, please use the endpoint `/auth/fab/v1` for this operation instead.*
      x-openapi-router-controller: airflow.api_connexion.endpoints.forward_to_fab_endpoint
      operationId: get_roles
      tags: [Role]
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageOffset"
        - $ref: "#/components/parameters/OrderBy"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleCollection"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"

    post:
      deprecated: true
      summary: Create a role
      description: |
        Create a new role.

        *This API endpoint is deprecated, please use the endpoint `/auth/fab/v1` for this operation instead.*
      x-openapi-router-controller: airflow.api_connexion.endpoints.forward_to_fab_endpoint
      operationId: post_role
      tags: [Role]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Role"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
  /roles/{role_name}:
    parameters:
      - $ref: "#/components/parameters/RoleName"

    get:
      deprecated: true
      summary: Get a role
      description: |
        Get a role.

        *This API endpoint is deprecated, please use the endpoint `/auth/fab/v1` for this operation instead.*
      x-openapi-router-controller: airflow.api_connexion.endpoints.forward_to_fab_endpoint
      operationId: get_role
      tags: [Role]
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      deprecated: true
      summary: Update a role
      description: |
        Update a role.

        *This API endpoint is deprecated, please use the endpoint `/auth/fab/v1` for this operation instead.*
      x-openapi-router-controller: airflow.api_connexion.endpoints.forward_to_fab_endpoint
      operationId: patch_role
      tags: [Role]
      parameters:
        - $ref: "#/components/parameters/UpdateMask"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Role"

      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      deprecated: true
      summary: Delete a role
      description: |
        Delete a role.

        *This API endpoint is deprecated, please use the endpoint `/auth/fab/v1` for this operation instead.*
      x-openapi-router-controller: airflow.api_connexion.endpoints.forward_to_fab_endpoint
      operationId: delete_role
      tags: [Role]
      responses:
        "204":
          description: Success.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
  /permissions:
    get:
      deprecated: true
      summary: List permissions
      description: |
        Get a list of permissions.

        *This API endpoint is deprecated, please use the endpoint `/auth/fab/v1` for this operation instead.*
      x-openapi-router-controller: airflow.api_connexion.endpoints.forward_to_fab_endpoint
      operationId: get_permissions
      tags: [Permission]
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageOffset"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionCollection"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/PermissionDenied"
...
components:
...
  schemas:
...
    Role:
      description: |
        a role item.

        *New in version 2.1.0*
      type: object
      properties:
        name:
          type: string
          description: |
            The name of the role

            *Changed in version 2.3.0*&#58; A minimum character length requirement ('minLength') is added.
          minLength: 1
        actions:
          type: array
          items:
            $ref: "#/components/schemas/ActionResource"
    RoleCollection:
      description: |
        A collection of roles.

        *New in version 2.1.0*
      type: object
      allOf:
        - type: object
          properties:
            roles:
              type: array
              items:
                $ref: "#/components/schemas/Role"
        - $ref: "#/components/schemas/CollectionInfo"
    Action:
      description: |
        An action Item.

        *New in version 2.1.0*
      type: object
      properties:
        name:
          type: string
          description: The name of the permission "action"
          nullable: false
    ActionCollection:
      description: |
        A collection of actions.

        *New in version 2.1.0*
      type: object
      allOf:
        - type: object
          properties:
            actions:
              type: array
              items:
                $ref: "#/components/schemas/Action"
        - $ref: "#/components/schemas/CollectionInfo"
    Resource:
      description: |
        A resource on which permissions are granted.

        *New in version 2.1.0*
      type: object
      properties:
        name:
          type: string
          description: The name of the resource
          nullable: false
    ActionResource:
      description: |
        The Action-Resource item.

        *New in version 2.1.0*
      type: object
      properties:
        action:
          type: object
          $ref: "#/components/schemas/Action"
          description: The permission action
        resource:
          type: object
          $ref: "#/components/schemas/Resource"
          description: The permission resource
...
    Error:
      description: |
        [RFC7807](https://tools.ietf.org/html/rfc7807) compliant response.
      type: object
      properties:
        type:
          type: string
          description: |
            A URI reference [RFC3986] that identifies the problem type. This specification
            encourages that, when dereferenced, it provide human-readable documentation for
            the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: number
          description: The HTTP status code generated by the API server for this occurrence of the problem.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
        instance:
          type: string
          description: |
            A URI reference that identifies the specific occurrence of the problem. It may or may
            not yield further information if dereferenced.
      required:
        - type
        - title
        - status
    CollectionInfo:
      description: Metadata about collection.
      type: object
      properties:
        total_entries:
          type: integer
          description: |
            Count of total objects in the current result set before pagination parameters
            (limit, offset) are applied.

    # Enums
...
  parameters:
...
    PageOffset:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        minimum: 0
      description: The number of items to skip before starting to collect the result set.
    PageLimit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        default: 100
      description: The numbers of items to return.

    # Database entity fields
...
    RoleName:
      in: path
      name: role_name
      schema:
        type: string
      required: true
      description: The role name
...
    OrderBy:
      in: query
      name: order_by
      schema:
        type: string
      required: false
      description: |
        The name of the field to order the results by.
        Prefix a field name with `-` to reverse the sort order.

        *New in version 2.1.0*
...
    UpdateMask:
      in: query
      name: update_mask
      schema:
        type: array
        items:
          type: string
      description: |
        The fields to update on the resource. If absent or empty, all modifiable fields are updated.
        A comma-separated list of fully qualified names of fields.
      style: form
      explode: false
...
  responses:
...
    "BadRequest":
      description: Client specified an invalid argument.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    # 401
    "Unauthenticated":
      description: Request not authenticated due to missing, invalid, authentication info.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    # 403
    "PermissionDenied":
      description: Client does not have sufficient permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    # 404
    "NotFound":
      description: A specified resource is not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    # 405
...
  securitySchemes:
    Basic:
      type: http
      scheme: basic
    GoogleOpenId:
      type: openIdConnect
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration
    Kerberos:
      type: http
      scheme: negotiate

# The API will provide support for plugins to support various authorization mechanisms.
# Detailed information will be available in the plugin specification.
security: []
...